Finish pass of consumer inline documentation; small updates to b2b
