+++ sdk/src/main/java/com/stytch/sdk/b2b/sessions/B2BSessions.kt
@@ -21,7 +21,7 @@ public interface B2BSessions {
 
     /**
      * Data class used for wrapping parameters used with Sessions authentication
-     * @param sessionDurationMinutes indicates how long the session should last before it expires
+     * @property sessionDurationMinutes indicates how long the session should last before it expires
      */
     public data class AuthParams(
         val sessionDurationMinutes: UInt? = null,
