+++ sdk/src/main/java/com/stytch/sdk/consumer/biometrics/Biometrics.kt
@@ -5,15 +5,19 @@ import com.stytch.sdk.common.Constants
 import com.stytch.sdk.consumer.BiometricsAuthResponse
 
 /**
- * Biometrics interface that encompasses registration and authentication functions
+ * The Biometrics interface provides methods for detecting biometric availability, registering, authenticating, and
+ * removing biometrics identifiers.
  */
 public interface Biometrics {
     /**
      * Data class used for wrapping parameters used with Biometrics registration flow
-     * @param context is the calling FragmentActivity
-     * @param sessionDurationMinutes indicates how long the session should last before it expires
-     * @param allowFallbackToCleartext opts-in to potentially unsafe behavior
-     * @param promptData is an optional biometric prompt configuration. If one is not provided a default will be created
+     * @property context is the calling FragmentActivity
+     * @property sessionDurationMinutes indicates how long the session should last before it expires
+     * @property allowFallbackToCleartext opts-in to potentially unsafe behavior
+     * @property promptData is an optional biometric prompt configuration. If one is not provided a default will be
+     * created
+     * @property allowDeviceCredentials opts-in to allowing the use of non-biometric device credentials (PIN, Pattern)
+     * as a fallback (on Android versions greater than Q)
      */
     public data class RegisterParameters(
         val context: FragmentActivity,
@@ -25,9 +29,10 @@ public interface Biometrics {
 
     /**
      * Data class used for wrapping parameters used with Biometrics authentication flow
-     * @param context is the calling FragmentActivity
-     * @param sessionDurationMinutes indicates how long the session should last before it expires
-     * @param promptData is an optional biometric prompt configuration. If one is not provided a default will be created
+     * @property context is the calling FragmentActivity
+     * @property sessionDurationMinutes indicates how long the session should last before it expires
+     * @property promptData is an optional biometric prompt configuration. If one is not provided a default will be
+     * created
      */
     public data class AuthenticateParameters(
         val context: FragmentActivity,
@@ -37,6 +42,9 @@ public interface Biometrics {
 
     /**
      * Data class used for wrapping parameters used to create a biometric prompt
+     * @property title The title to be displayed in the Biometric Prompt
+     * @property subTitle The subtitle to be displayed in the Biometric Prompt
+     * @property negativeButtonText "Cancel" button text to display, if device credentials are disallowed
      */
     public data class PromptData(
         val title: String,
@@ -46,11 +54,17 @@ public interface Biometrics {
 
     /**
      * Indicates if there is an existing biometric registration on device.
+     * @param context is the calling FragmentActivity
+     * @return Boolean
      */
     public fun isRegistrationAvailable(context: FragmentActivity): Boolean
 
     /**
      * Indicates if the biometric sensor is available, and provides a reasoning if not
+     * @param context is the calling FragmentActivity
+     * @param allowDeviceCredentials whether or not you choose to allow the use of non-biometric device credentials
+     * (PIN, Pattern) as a fallback (on Android versions greater than Q)
+     * @return [BiometricAvailability]
      */
     public fun areBiometricsAvailable(
         context: FragmentActivity,
@@ -59,16 +73,19 @@ public interface Biometrics {
 
     /**
      * Clears existing biometric registrations stored on device. Useful when removing a user from a given device.
+     * @return Boolean
      */
     public suspend fun removeRegistration(): Boolean
 
     /**
      * Clears existing biometric registrations stored on device. Useful when removing a user from a given device.
+     * @param callback A callback that receives a Boolean
      */
     public fun removeRegistration(callback: (Boolean) -> Unit)
 
     /**
      * Indicates if the device is using the Android KeyStore
+     * @return Boolean
      */
     public fun isUsingKeystore(): Boolean
 
@@ -77,6 +94,7 @@ public interface Biometrics {
      * The user will later be able to start a new session with biometrics or use biometrics as an additional
      * authentication factor.
      * @param parameters required to register a biometrics key
+     * @return [BiometricsAuthResponse]
      */
     public suspend fun register(parameters: RegisterParameters): BiometricsAuthResponse
 
@@ -85,7 +103,7 @@ public interface Biometrics {
      * The user will later be able to start a new session with biometrics or use biometrics as an additional
      * authentication factor.
      * @param parameters required to register a biometrics key
-     * @param callback calls callback with BiometricsAuthResponse response from backend
+     * @param callback a callback that receives a [BiometricsAuthResponse]
      */
     public fun register(
         parameters: RegisterParameters,
@@ -97,6 +115,7 @@ public interface Biometrics {
      * biometric reader and returns an updated session object by either starting a new session or adding the biometric
      * factor to an existing session.
      * @param parameters required to authenticate a biometrics key
+     * @return [BiometricsAuthResponse]
      */
     public suspend fun authenticate(parameters: AuthenticateParameters): BiometricsAuthResponse
 
@@ -105,7 +124,7 @@ public interface Biometrics {
      * biometric reader and returns an updated session object by either starting a new session or adding the biometric
      * factor to an existing session.
      * @param parameters required to authenticate a biometrics key
-     * @param callback calls callback with BiometricsAuthResponse response from backend
+     * @param callback a callback that receives a [BiometricsAuthResponse]
      */
     public fun authenticate(
         parameters: AuthenticateParameters,
