+++ sdk/src/main/java/com/stytch/sdk/consumer/userManagement/UserManagement.kt
@@ -7,20 +7,12 @@ import com.stytch.sdk.consumer.network.models.UserData
 /**
  * The UserManagement interface provides methods for retrieving an authenticated user and deleting authentication
  * factors from an authenticated user.
- *
- * You can choose to get the local representation of the user, without making a network request, with the
- * StytchClient.user.getSyncUser() method.
- *
- * If you want to get the freshest representation of the user from the Stytch servers, use the
- * StytchClient.user.getUser() method.
- *
- * To remove an authentication factor from a user, use the StytchClient.user.deleteFactor() method.
  */
 public interface UserManagement {
 
     /**
      * Fetches a user from the current session
-     * @return [UserResponse] response from backend
+     * @return [UserResponse]
      */
     public suspend fun getUser(): UserResponse
 
@@ -38,7 +30,7 @@ public interface UserManagement {
 
     /**
      * Deletes a [UserAuthenticationFactor] from the currently authenticated user
-     * @return [DeleteFactorResponse] response from backend
+     * @return [DeleteFactorResponse]
      */
     public suspend fun deleteFactor(factor: UserAuthenticationFactor): DeleteFactorResponse
